requirements.txt
setuptools>=0.9.8
colorama>=0.3.9
mysql-connector-python>=2.2.2
